Cast some more to keep both 32 bit and 64 bit simulator platforms happy.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@12819 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs/aiff.c
@@ -99,7 +99,8 @@ next_track:
         i = ((buf[4]<<24)|(buf[5]<<16)|(buf[6]<<8)|buf[7]);
         if (memcmp(buf, "COMM", 4) == 0) {
             if (i != 18) {
-                DEBUGF("CODEC_ERROR: 'COMM' chunk size=%lu != 18\n", i);
+                DEBUGF("CODEC_ERROR: 'COMM' chunk size=%lu != 18\n",
+                       (unsigned long)i);
                 i = CODEC_ERROR;
                 goto done;
             }
@@ -136,7 +137,7 @@ next_track:
             i = 8 + offset2snd; /* advance to the beginning of data */
         } else {
             DEBUGF("unsupported AIFF chunk: '%c%c%c%c', size=%lu\n",
-                   buf[0], buf[1], buf[2], buf[3], i);
+                   buf[0], buf[1], buf[2], buf[3], (unsigned long)i);
         }
 
         if (i & 0x01) /* odd chunk sizes must be padded */
